--- Code/KvalitetesLedelsesSystem/.idea/.idea.KvalitetesLedelsesSystem/.idea/.gitignore
@@ -1,13 +1,0 @@
-﻿# Default ignored files
-/shelf/
-/workspace.xml
-# Rider ignored files
-/projectSettingsUpdater.xml
-/.idea.KvalitetesLedelsesSystem.iml
-/contentModel.xml
-/modules.xml
-# Editor-based HTTP Client requests
-/httpRequests/
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
