+++ Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/App.xaml
@@ -4,11 +4,6 @@
              xmlns:local="clr-namespace:KvalitetesLedelsesSystem"
              StartupUri="Views\MainWindow.xaml">
     <Application.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="pack://application:,,,/
-                                    Haley.WPF;component/Dictionaries/haleyRD.xaml"/>
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
+         
     </Application.Resources>
 </Application>
