--- Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/Models/Colour.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace KvalitetesLedelsesSystem.Models
-{
-    public class Colour
-    {
-        
-        public string selectedColour;
-
-        public Colour(string BaseColour)
-        {
-            selectedColour = BaseColour;
-        }
-    }
-}
