--- Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/ViewModels/Commands/AdminMenuCommand.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Input;
-
-namespace KvalitetesLedelsesSystem.ViewModels.Commands
-{
-    internal class AdminMenuCommand : ICommand
-    {
-        public event EventHandler? CanExecuteChanged;
-
-        public bool CanExecute(object? parameter)
-        {
-            return true;
-        }
-
-        public void Execute(object? parameter)
-        {
-            if (parameter is MainViewModel mvm)
-            {
-                var mainWindow = Application.Current.MainWindow as MainWindow;
-                mainWindow?.NavigateTo_AdminMenu();
-            }
-        }
-    }
-}
