--- Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/ViewModels/Commands/PlanCommand.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Input;
-
-
-namespace KvalitetesLedelsesSystem.ViewModels.Commands
-{
-    public class PlanCommand : ICommand
-    {
-        public event EventHandler? CanExecuteChanged;
-
-        public bool CanExecute(object? parameter)
-        {
-            return true;
-        }
-
-        public void Execute(object? parameter)
-        {
-            var mainWindow = Application.Current.MainWindow as MainWindow;
-            mainWindow?.NavigateTo_Plan();
-        }
-    }
-}
