+++ Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/Views/Admin/AdminCRUD.xaml
@@ -6,17 +6,13 @@
         xmlns:local="clr-namespace:KvalitetesLedelsesSystem.Views"
         mc:Ignorable="d"
         Title="AdminCRUD" Height="450" Width="800">
-    <Grid Background="{Binding Path=colourVMs[1].SelectedColour}">
-        <Grid.RowDefinitions >
-
+    <Grid Background="#3F3F3F">
+        <Grid.RowDefinitions>
             <RowDefinition Height="12*"/>
             <RowDefinition Height="*"/>
         </Grid.RowDefinitions>
 
-
-
-        <Grid Background="{Binding Path=colourVMs[0].SelectedColour}" Grid.Row="0">
-
+        <Grid Background="#FF1E1E1E" Grid.Row="0">
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="*"/>
                 <ColumnDefinition Width="2*"/>
