+++ Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/Views/Admin/AdminLogin.xaml.cs
@@ -1,5 +1,4 @@
-﻿using KvalitetesLedelsesSystem.ViewModels;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -21,20 +20,10 @@ namespace KvalitetesLedelsesSystem.Views
     /// </summary>
     public partial class AdminLogin : Page
     {
-        MainViewModel mvm = new MainViewModel();
+        
         public AdminLogin()
         {
             InitializeComponent();
-            DataContext = mvm;
-        }
-       
-        private void Button_Click(object sender, RoutedEventArgs e)
-        {
-            
-            {
-                NavigationService.GoBack();
-            }
-
         }
     }
 }
