+++ Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/Views/Admin/AdminMenu.xaml
@@ -6,11 +6,9 @@
       xmlns:local="clr-namespace:KvalitetesLedelsesSystem.Views"
       mc:Ignorable="d" 
       d:DesignHeight="450" d:DesignWidth="800"
-      xmlns:hw ="http://schemas.hpod9.com/haley/wpf"
-      xmlns:hm="http://schemas.hpod9.com/haley/mvvm"
       Title="AdminMenu">
 
-    <Grid Background="{Binding Path=colourVMs[1].SelectedColour}">
+    <Grid Background="#3F3F3F">
         <Grid.RowDefinitions >
             <RowDefinition Height="*"/>
             <RowDefinition Height="12*"/>
@@ -24,22 +22,21 @@
                 <ColumnDefinition Width="*"/>
             </Grid.ColumnDefinitions>
 
-            <Button FontSize="20" FontWeight="Bold" Content="⮐" Background="{Binding Path=colourVMs[1].SelectedColour}" Foreground="#FFFFFFFF" Grid.Column="0" Click="Button_Click_1"/>
+            <Button FontSize="20" FontWeight="Bold" Content="⮐" Background="#3F3F3F" Foreground="#FFFFFFFF" Grid.Column="0"/>
             <TextBlock Text="Admin Menu" Foreground="#FFFFFFFF" FontSize="24" FontWeight="Bold"  Grid.Column="1" VerticalAlignment="Center" HorizontalAlignment="Center"/>
-            <Image Source="{Binding imageVMs[0].SelectedImage}"  Grid.Column="2"/>
+            <Image Source="/Views/Societate transparent.png"  Grid.Column="2"/>
 
         </Grid>
 
-        <Grid Background="{Binding Path=colourVMs[0].SelectedColour}" Grid.Row="1">
+        <Grid Background="#FF1E1E1E" Grid.Row="1">
 
             <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="3*"/>
-                <ColumnDefinition/>
-                <ColumnDefinition Width="8*"/>
-                <ColumnDefinition Width="4*"/>
+                <ColumnDefinition Width="*"/>
+                <ColumnDefinition Width="2*"/>
+                <ColumnDefinition Width="*"/>
             </Grid.ColumnDefinitions>
 
-            <Grid Grid.Column="2">
+            <Grid Grid.Column="1">
                 <Grid.RowDefinitions>
                     <RowDefinition Height="*"/>
                     <RowDefinition Height="2*"/>
@@ -48,18 +45,15 @@
                     <RowDefinition Height="*"/>
                     <RowDefinition Height="5*"/>
                     <RowDefinition Height="*"/>
-                    <RowDefinition Height="*"/>
-
                 </Grid.RowDefinitions>
-                <Button Name="Default" Grid.Row="6" Grid.Column="0" Content="Default Settings" Command="{Binding Path=DefautSettingsCommand }" CommandParameter="Default"/>
+
                 <Separator Grid.Row="0"/>
                 <Separator Grid.Row="2"/>
                 <Separator Grid.Row="4"/>
-                <Separator Grid.Row="7" Margin="0,1,0,3"/>
+                <Separator Grid.Row="6"/>
 
                 <Grid Grid.Row="5">
                     <Grid.RowDefinitions>
-                        <RowDefinition Height="*"/>
                         <RowDefinition Height="*"/>
                         <RowDefinition Height="*"/>
                         <RowDefinition Height="*"/>
@@ -77,10 +71,10 @@
                     <Label Content="Change Foreground Color" FontSize="13" FontWeight="Bold" Foreground="White" Grid.Column="2" Grid.Row="0" HorizontalAlignment="Left" VerticalAlignment="Bottom"/>
                     <Label Content="Change Background Color" FontSize="13" FontWeight="Bold" Foreground="White" Grid.Column="0" Grid.Row="3" HorizontalAlignment="Left" VerticalAlignment="Bottom"/>
                     <Label Content="Change Accent Color" FontSize="13" FontWeight="Bold" Foreground="White" Grid.Column="2" Grid.Row="3" HorizontalAlignment="Left" VerticalAlignment="Bottom"/>
-                    <Button Command="{Binding Path=UpdatePath}" CommandParameter="Logo" Content="Open File" FontSize="13" Grid.Column="0" Grid.Row="1"/>
-                    <Button x:Name="Foreground" Content="Select Foreground Color" FontSize="13" Grid.Column="2" Grid.Row="1"  Command="{Binding Path=UpdateColor}" CommandParameter="Foreground"/>
-                    <Button x:Name="BackgroundColour" Content="Select Background Color" FontSize="13" Grid.Column="0" Grid.Row="4" Command="{Binding Path=UpdateColor}" CommandParameter="Background" />
-                    <Button x:Name="Accent" Content="Select Accent Color" FontSize="13" Grid.Column="2" Grid.Row="4" Command="{Binding Path=UpdateColor}" CommandParameter="Accent"/>
+                    <Button Content="Open File" FontSize="13" Grid.Column="0" Grid.Row="1" Click="Button_Click"/>
+                    <Button Content="Select Foreground Color" FontSize="13" Grid.Column="2" Grid.Row="1"/>
+                    <Button Content="Select Background Color" FontSize="13" Grid.Column="0" Grid.Row="4"/>
+                    <Button Content="Select Accent Color" FontSize="13" Grid.Column="2" Grid.Row="4"/>
 
                 </Grid>
 
@@ -97,7 +91,7 @@
                     </Grid.ColumnDefinitions>
 
                     <Label Content="User Settings" FontSize="13" FontWeight="Bold" Foreground="White" Grid.Column="0" Grid.ColumnSpan="3" Grid.Row="0" HorizontalAlignment="Left" VerticalAlignment="Bottom"/>
-                    <Button Content="Open User Settings" Command="{Binding Path=PersonListCommand}" CommandParameter="{Binding}" FontSize="13" Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="3"/>
+                    <Button Content="Open User Settings" FontSize="13" Grid.Column="0" Grid.Row="1" Grid.ColumnSpan="3"/>
                 </Grid>
 
 
@@ -115,8 +109,8 @@
 
                     <Label Content="Contingency Plan" FontSize="13" FontWeight="Bold" Foreground="White" Grid.Column="0" Grid.Row="0" HorizontalAlignment="Left" VerticalAlignment="Bottom"/>
                     <Label Content="Contingency Drawing" FontSize="13" FontWeight="Bold" Foreground="White" Grid.Column="2" Grid.Row="0" HorizontalAlignment="Left" VerticalAlignment="Bottom"/>
-                    <Button Command="{Binding Path=UpdatePath}" CommandParameter="ContingencyPlan" x:Name ="ContingencyPlanBtn" FontSize="13" Content="Open File" Grid.Column="0" Grid.Row="1"/>
-                    <Button Command="{Binding Path=UpdatePath}" CommandParameter="ContingencyDrawing" x:Name="ContingencyDrawingbtn" FontSize="13" Content="Open File" Grid.Column="2" Grid.Row="1" />
+                    <Button FontSize="13" Content="Open File" Grid.Column="0" Grid.Row="1"/>
+                    <Button FontSize="13" Content="Open File" Grid.Column="2" Grid.Row="1"/>
                 </Grid>
 
             </Grid>
