+++ Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/Views/Admin/PersonList.xaml.cs
@@ -27,10 +27,5 @@ namespace KvalitetesLedelsesSystem.Views
             InitializeComponent();
             DataContext = mvm;
         }
-
-        private void Button_Click(object sender, RoutedEventArgs e)
-        {
-            NavigationService.GoBack();
-        }
     }
 }
