+++ Code/KvalitetesLedelsesSystem/KvalitetesLedelsesSystem/Views/Plan.xaml
@@ -8,22 +8,22 @@
       d:DesignHeight="450" d:DesignWidth="800"
       Title="Plan">
 
-    <Grid Background="{Binding Path=colourVMs[0].SelectedColour}">
+    <Grid Background="#FF1E1E1E">
         <Grid.RowDefinitions >
             <RowDefinition Height="*"/>
             <RowDefinition Height="12*"/>
             <RowDefinition Height="*" />
         </Grid.RowDefinitions>
         <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="4*"/>
-            <ColumnDefinition Width="4*"/>
+            <ColumnDefinition Width="*"/>
+            <ColumnDefinition Width="5*"/>
             <ColumnDefinition Width="*"/>
         </Grid.ColumnDefinitions>
-
-        <Button x:Name="CheckIn_ud_Button" Command="{Binding CheckCommand}" CommandParameter="{Binding}" FontSize="16" FontWeight="Bold" Content="Check In/Ud" Background="{Binding Path=colourVMs[1].SelectedColour}" Foreground="#FFFFFFFF" Grid.Column="0"/>
-        <Button x:Name="ContingencyPlan_Button" FontSize="16" FontWeight="Bold" Content="Contingency Plan" Background="{Binding Path=colourVMs[1].SelectedColour}" Foreground="#FFFFFFFF" Grid.Column="1"/>
-        <Button FontSize="20"  Content="⚙" Command="{Binding Path = AdminLogInCommand}" CommandParameter="{Binding}" Background="{Binding Path=colourVMs[1].SelectedColour}" Foreground="#FFFFFFFF" Grid.Column="2"/>
-        <Border Grid.Row="2" Grid.ColumnSpan ="3" Background="{Binding Path=colourVMs[1].SelectedColour}" />
+        
+        <Button x:Name="CheckIn_ud_Button" FontSize="16" FontWeight="Bold" Content="Check In/Ud" Background="#3F3F3F" Foreground="#FFFFFFFF" Grid.Column="0" Click="CheckinUd_Click"/>
+        <Button x:Name="ContingencyPlan_Button" FontSize="16" FontWeight="Bold" Content="Contingency Plan" Background="#3F3F3F" Foreground="#FFFFFFFF" Grid.Column="1"/>
+        <Button FontSize="20"  Content="⚙" Background="#3F3F3F" Foreground="#FFFFFFFF" Grid.Column="2"/>
+        <Border Grid.Row="2" Grid.ColumnSpan ="3" Background="#3F3F3F" />
        
         <Grid Grid.Row="1" Grid.ColumnSpan="3">
             <Grid.ColumnDefinitions>
@@ -32,11 +32,11 @@
                 <ColumnDefinition Width="*"/>
                 <ColumnDefinition Width="20*"/>
                 <ColumnDefinition Width="*"/>
-            </Grid.ColumnDefinitions>      
+            </Grid.ColumnDefinitions>   
 
             <WebBrowser x:Name="pdfWebViewer" Grid.Column="3" ></WebBrowser>
 
-            <Image x:Name="ContingencyDrawing" Source="{Binding imageVMs[1].SelectedImage}" Grid.Row="1" Grid.Column="1" />
+            <Image Grid.Row="1" Grid.Column="1"  Source="/Views/Societate transparent.png"/>
         </Grid>
 
     </Grid>
