--- Diagrams/CheckSD.plantuml
@@ -1,29 +1,0 @@
-@startuml test
-skinparam style strictuml
-skinparam sequenceMessageAlign center
-title Check
-
-Actor "User/Check" as user
-
-participant ":CheckInCommand" as CIC
-participant ":MainViewModel" as MVM
-participant ":UserViewModel" as UVM
-participant ":UserRepository" as UR
-
-user -> CIC ++: Execute()
-    
-    CIC -> MVM ++: UserCheck(UserName)      
-
-        MVM -> UVM ++: UserName == UserviewModel.UserName
-
-
-           
-        return Return True
-
-    MVM -> UVM ++: ChangeCheck(userRepository:UserRepository)
-    UVM -> UR ++: ChangeCheck()
-    UR -> UR : UpdateFile()
-    return
-return
-return
-return
